+++ pages/api/blogcategory/[id].js
@@ -153,7 +153,12 @@ export default async function handler(req, res) {
                   "../../../../../public/assets/upload/blog",
                   oldImage
                 );
-                unlink(oldImagePath);
+                if (fs.existsSync(oldImagePath)) {
+                  await unlink(oldImagePath);
+                } else {
+                  console.log(`File does not exist: ${oldImagePath}`);
+                }
+                
               }
 
               const oldPath = imageFile[0].filepath;
