+++ pages/api/products/productdrawing/perdrawingdata/[id].js
@@ -94,7 +94,11 @@ export default async function handler(req, res) {
           if (productDocs.length !== 0) {
             const oldImage = productDocs[0].pdf_link;
             const oldImagePath = path.join(projectDirectory, oldImage);
-            await unlink(oldImagePath);
+            if (fs.existsSync(oldImagePath)) {
+              await unlink(oldImagePath);
+            } else {
+              console.log(`File does not exist: ${oldImagePath}`);
+            }
           }
         }
