+++ pages/api/products/productimages/perimgdata/[id].js
@@ -117,7 +117,11 @@ export default async function handler(req, res) {
           if (productImg.length !== 0) {
             const oldImage = productImg[0].product_image;
             const oldImagePath = path.join(projectDirectory, oldImage);
-            await unlink(oldImagePath);
+            if (fs.existsSync(oldImagePath)) {
+              await unlink(oldImagePath);
+            } else {
+              console.log(`File does not exist: ${oldImagePath}`);
+            }
           }
         }
