+++ pages/api/testimonial/[id].js
@@ -111,7 +111,11 @@ export default async function handler(req, res) {
           if (testimonial.length !== 0) {
             const oldImage = testimonial[0].testimonial_image;
             const oldImagePath = path.join(projectDirectory, oldImage);
-            await unlink(oldImagePath);
+            if (fs.existsSync(oldImagePath)) {
+              await unlink(oldImagePath);
+            } else {
+              console.log(`File does not exist: ${oldImagePath}`);
+            }
           }
         }
 
@@ -148,7 +152,11 @@ export default async function handler(req, res) {
         );
         const newPath = path.join(projectDirectory, testimonialImage);
         console.log(newPath);
-        await unlink(newPath);
+        if (fs.existsSync(newPath)) {
+          await unlink(newPath);
+        } else {
+          console.log(`File does not exist: ${newPath}`);
+        }
       }
 
       // Process the data and send the response
